Move top level logic into new `ukcli` package
